Revert unneeded changes from "Fixes for nested multipliers (#59)"
This partially reverts commit 39725d333355b2a1ef21dac3948909cf1e5d4744,
only keeping the part relevant for nested support and a PHPUnit deprecation fix.

The commit introduced a regression in `testGroupManualRenderWithButtons`
so this will minimize the changes that could affect it.

- Latte 2 macros are trivial to migrate and unclearly named, no need to keep them for BC.
- The onSuccess handlers were already fixed in ed96ba0276a2db878841836c403f60568c7a3a98.
- `testGroupManualRenderWithButtons` look like remnants of debugging effort, they do not fix the test.

For reference, the reverted change is the following commit that was squashed
into 39725d333355b2a1ef21dac3948909cf1e5d4744, except for the change to `testSendNested`:
https://github.com/contributte/forms-multiplier/pull/59/commits/2c33de22b8343ba08210764711e4776a5bf58227