Mostly revert "Fixes for nested multipliers (#59)"
This partially reverts commit 39725d333355b2a1ef21dac3948909cf1e5d4744.

The commit introduced a regression in `testGroupManualRenderWithButtons`,
changed the internal architecture in a way that made it harder to reason about
(the Multiplier inconsistently took over some responsibilities of ComponentResolver).
Additionally, it applied the following changes that are unrelated
to the purported fix and not really necessary:

- Latte 2 macros are trivial to migrate and unclearly named, no need to keep them for BC.
- The onSuccess handlers were already fixed in ed96ba0276a2db878841836c403f60568c7a3a98.
- `testGroupManualRenderWithButtons` look like remnants of debugging effort, they do not fix the test.

We are only keeping the following:

- Test for nested support (to be implemented later).
- Fix for `assertMatchesRegularExpression` deprecation in PHPUnit.

For reference, the reverted change is the following commit that was squashed
into 39725d333355b2a1ef21dac3948909cf1e5d4744, except for the change to `testSendNested`:
https://github.com/contributte/forms-multiplier/pull/59/commits/2c33de22b8343ba08210764711e4776a5bf58227